update
overhauled the wind particle effects.
added a small trail on both ends of the snowboard when the player is in the air.
added a wind particle effect that plays when the player falls rapidly
adjusted the speed wind particle effect to fill the screen more.
added wind burst particle effect when the player hits a speed boost.