update
Fixed bug in which playing as the Bear would bearly move left and right
Fixed bug in which the player can acess the pause menu when crossing the finish line.
Fixed bug in which the player can access the credits scene through scrolling the pause menu using the arrow keys to access the hidden finish line button.
Updated the trail to stay on the ground when the player jumps.