BEEG bug fixes
Fixed Leaderboard UI not showing the data when finishing a race (Needs more Testing!)
Fixed a BIG bug in which the gameloop controller was being duplicated each time you enter the main menu
Fixed the bug in which the gameloop controller was freaking out when the player hops back and forth between levels. This was the real major one we talked about last meeting.
Fixed a Visual bug in which loading into Level 1 the text displayed on the loading screen showed "Credits". Same for applies for all levels now.
Fixed visual artifacts on the boost panels
added more pattern assets